Remove InstancePatchFix patches (fix reverse patch crash)
